move registration and changes app code to eventstore
