+++ .github/workflows/test.yaml
@@ -40,7 +40,7 @@ jobs:
       - name: Lint
         run: |
           flake8
-          python manage.py makemigrations eventstore --dry-run | grep 'No changes detected' || (echo 'There are changes which require migrations.' && exit 1)
+          python manage.py makemigrations registrations changes eventstore --dry-run | grep 'No changes detected' || (echo 'There are changes which require migrations.' && exit 1)
           black --check .
           isort -c -rc .
       - name: Test
