+++ eventstore/test_tasks.py
@@ -91,7 +91,7 @@ class HandleOldWaitingForHelpdeskContactsTests(TestCase):
         )
 
         responses.add(
-            responses.POST, f"{'http://turn/v1/chats/27820001001/archive'}", json={}
+            responses.POST, "http://turn/v1/chats/27820001001/archive", json={}
         )
 
         tasks.handle_expired_helpdesk_contacts()
@@ -111,7 +111,7 @@ class HandleOldWaitingForHelpdeskContactsTests(TestCase):
             json.loads(turn_archive.request.body),
             {
                 "before": "ABGGJ4NjeFMfAgo-sCqKaSQU4UzP",
-                "reason": f"{'Auto archived after 11 days'}",
+                "reason": "Auto archived after 11 days",
             },
         )
 
@@ -700,7 +700,8 @@ class ProcessAdaAssessmentNotificationTests(TestCase):
         self.assertEqual(triage.risk, Covid19Triage.RISK_MODERATE),
         self.assertEqual(triage.gender, Covid19Triage.GENDER_FEMALE),
         self.assertEqual(
-            triage.completed_timestamp, datetime.datetime(2021, 1, 2, 3, 4, 5)
+            triage.completed_timestamp,
+            datetime.datetime(2021, 1, 2, 3, 4, 5, tzinfo=datetime.timezone.utc),
         )
         self.assertEqual(triage.created_by, "test"),
         self.assertEqual(triage.data, {"age": 30, "pregnant": False}),
