+++ ndoh_hub/settings.py
@@ -59,7 +59,7 @@ INSTALLED_APPS = (
     "django_prometheus",
     # us
     "registrations",
-    "changes",
+    # "changes",
     "eventstore",
     "ada",
 )
@@ -139,7 +139,7 @@ USE_I18N = True
 
 USE_L10N = True
 
-# USE_TZ = True
+USE_TZ = True
 
 LANGUAGES = [
     ("afr-za", "Afrikaans"),
