+++ ndoh_hub/urls.py
@@ -37,7 +37,6 @@ from eventstore.views import (
     ResearchOptinSwitchViewSet,
 )
 from ndoh_hub.decorators import internal_only
-from registrations import views
 
 admin.site.site_header = os.environ.get("HUB_TITLE", "NDOH Hub Admin")
 
@@ -82,15 +81,7 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     url(r"^api/auth/", include("rest_framework.urls", namespace="rest_framework")),
     url(r"^api/token-auth/", obtain_auth_token),
-    url(r"^api/metrics/", views.MetricsView.as_view()),
-    url(
-        r"^api/health/jembi-facility/",
-        views.JembiFacilityCheckHealthcheckView.as_view(),
-    ),
-    url(r"^api/health/", views.HealthcheckView.as_view()),
     url(r"^docs/", include_docs_urls(title="NDOH Hub")),
-    url(r"^", include("registrations.urls")),
-    url(r"^", include("changes.urls")),
     path("", include("ada.urls")),
     url(r"^api/v1/forgetcontact/", ForgetContactView.as_view(), name="forgetcontact"),
     path("api/v2/", include(v2router.urls)),
